yes capitalize but depends on conference
Capitalize the first and last words of the title, no matter what they are.
Capitalize all major words in the title, including nouns, pronouns, verbs, adjectives, adverbs, and some conjunctions.
Lowercase articles (a, an, the), coordinating conjunctions (and, but, or, for, nor), and prepositions, regardless of length, unless they are the first or last word of the title.
Lowercase the 'to' in an infinitive.
So your title would be transformed to: